Use appropriate PHP executable with both setup and tests.
It looks like the test runner and WordPress specify the PHP version to use to install WordPress before the tests, but this doesn't match with the version of PHP used to run the tests themselves.

This resulted in the test runner reporting a certain version of PHP being used to run the tests, when it was really the current default PHP in the env.

This commit attempts to remedy this.