Remove `npm` caching from the example GHA workflow.
This is difficult to set up because a `package-lock.json` file is required in order to create a cache key. In this context, the lock file should be the one from WordPress Core. But since WordPress is not checked out until the `prepare.php` step, caching cannot be configured before hand.