Change the default value of `WPT_FLAVOR`
This changes the default value for `WPT_FLAVOR` in the `.env.default` file from `1` (multisite) to `0` (single site).

While receiving test reports from environments running multisite offers additional confirmation, running the tests as a single site is the more appropriate default.

The code throughout the repository responsible for handling the `WPT_FLAVOR` environment variable currently uses single site as the fallback when the variable is not set.

Co-Authored-By: Javier Casares <javier@casares.org>